Move env vars to their own modules
