+++ gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/common/BasePlugin.kt
@@ -5,6 +5,8 @@ import org.gradle.api.Project
 import org.gradle.api.file.Directory
 import org.gradle.api.provider.Provider
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+import java.io.FileInputStream
+import java.util.*
 
 @Suppress("unused")
 class BasePlugin : Plugin<Project> {
@@ -21,3 +23,10 @@ class BasePlugin : Plugin<Project> {
 }
 
 fun Project.binaryDestination(): Provider<Directory> = rootProject.layout.buildDirectory.dir("services/${project.name}")
+
+fun Project.envVar(key: String): String {
+    val propFile = file("./env.properties")
+    val properties = Properties()
+    properties.load(FileInputStream(propFile))
+    return properties.getProperty(key)
+}
