+++ search-data/src/commonMain/kotlin/com/gchristov/thecodinglove/searchdata/SearchDataModule.kt
@@ -1,7 +1,6 @@
 package com.gchristov.thecodinglove.searchdata
 
 import com.gchristov.thecodinglove.commonfirebasedata.FirebaseAdmin
-import com.gchristov.thecodinglove.commonkotlin.BuildConfig
 import com.gchristov.thecodinglove.commonkotlin.JsonSerializer
 import com.gchristov.thecodinglove.commonkotlin.di.DiModule
 import com.gchristov.thecodinglove.commonnetwork.NetworkClient
